refactor: Move helper function to common file
